+++ api/expense-manager/src/main/java/com/example/expense_manager/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.example.expense_manager;
+package com.example.expense_manager.repository;
 
+import com.example.expense_manager.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
