--- api/expense-manager/src/main/resources/static/loginHTML.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Login</title>
-</head>
-<body>
-<h2>Login</h2>
-<form action="/login" method="post">
-    <div>
-        <label>Username:</label>
-        <input type="text" name="username" />
-    </div>
-    <div>
-        <label>Password:</label>
-        <input type="password" name="password" />
-    </div>
-    <div>
-        <button type="submit">Login</button>
-    </div>
-</form>
-</body>
-</html>
