--- api/expense-manager/src/main/resources/static/registerHTML.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Register</title>
-</head>
-<body>
-<h2>Register</h2>
-<form action="/register" method="post">
-    <div>
-        <label>Name:</label>
-        <input type="text" name="name" required />
-    </div>
-    <div>
-        <label>Username:</label>
-        <input type="text" name="username" required />
-    </div>
-    <div>
-        <label>Password:</label>
-        <input type="password" name="password" required />
-    </div>
-    <div>
-        <button type="submit">Register</button>
-    </div>
-</form>
-</body>
-</html>
